Melhed dev playerstate

+++ backend-hvz/src/main/java/com/example/backendhvz/services/game/GameServiceImpl.java
@@ -1,7 +1,10 @@
 package com.example.backendhvz.services.game;
 
+import com.example.backendhvz.enums.PlayerState;
 import com.example.backendhvz.models.Game;
+import com.example.backendhvz.models.Player;
 import com.example.backendhvz.repositories.GameRepository;
+import com.example.backendhvz.repositories.PlayerRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
@@ -10,9 +13,11 @@ import java.util.Collection;
 public class GameServiceImpl implements GameService {
 
     private final GameRepository gameRepository;
+    private final PlayerRepository playerRepository;
 
-    public GameServiceImpl(GameRepository gameRepository) {
+    public GameServiceImpl(GameRepository gameRepository, PlayerRepository playerRepository) {
         this.gameRepository = gameRepository;
+        this.playerRepository = playerRepository;
     }
 
     @Override
@@ -31,12 +36,33 @@ public class GameServiceImpl implements GameService {
     }
 
     @Override
+    public Game addGame(Game game, Long creatingPlayerId) {
+        Player creatingPlayer = playerRepository.findById(creatingPlayerId).get();
+        if(creatingPlayer.getState() != PlayerState.ADMINISTRATOR) return null;
+        return gameRepository.save(game);
+    }
+
+    @Override
     public Game update(Game game) {
         return gameRepository.save(game);
     }
 
     @Override
+    public Game updateGame(Game game, Long updatingPlayerId) {
+        Player updatingPlayer = playerRepository.findById(updatingPlayerId).get();
+        if(updatingPlayer.getState() != PlayerState.ADMINISTRATOR) return null;
+        return gameRepository.save(game);
+    }
+
+    @Override
     public void deleteById(Long gameId) {
+        gameRepository.deleteById(gameId);
+    }
+
+    @Override
+    public void deleteGameById(Long gameId, Long deletingPlayerId) {
+        Player deletingPlayer = playerRepository.findById(deletingPlayerId).get();
+        if(deletingPlayer.getState() != PlayerState.ADMINISTRATOR) return;
         gameRepository.deleteById(gameId);
     }
